chore: split formula utils and its tests into smaller files [PT-186452591]
